requirementsB.txt
fastapi
uvicorn
pydantic
requests
langchain-groq
langchain-huggingface
langchain-text-splitters
faiss-cpu
sentence-transformers
python-dotenv
yt-dlp